+++ .github/workflows/test.yml
@@ -9,7 +9,7 @@ env:
   PUBLIC_SENTRY_DSN:
   POSTGRES_HOST: postgres
   POSTGRES_PORT: 5432
-  DATABASE_URL: postgres://postgres:postgres@postgres:5432/oh-my-gtd?schema=public
+  DATABASE_URL: postgres://postgres:postgres@localhost:5432/oh-my-gtd?schema=public
   REDIS_URL:
   JWT_SESSION_SECRET: ${{ secrets.JWT_SESSION_SECRET }}
 
@@ -31,6 +31,8 @@ jobs:
           --health-interval 10s
           --health-timeout 5s
           --health-retries 5
+        ports:
+          - 5432:5432
 
     steps:
       - name: Checkout code
