+++ src/routes/+page.server.ts
@@ -1,6 +1,6 @@
 import { fail, redirect } from '@sveltejs/kit';
 import type { Actions } from './$types';
-import { signup } from '$lib/services/user';
+import { signup } from '$lib/services/auth';
 import { ApiError } from '$lib/utils/api-error';
 import { ApiResponse } from '$lib/utils/api-response';
 import { Session } from '$lib/services/session';
