+++ src/routes/dashboard/+page.server.ts
@@ -4,6 +4,6 @@ import { Logger } from '$lib/services/logger';
 
 export const load: PageServerLoad = async ({ locals }) => {
     if (!locals.session.user) redirect(303, '/?signin=true');
-
-	Logger.info('locals: ', locals.session.user);
+    
+    Logger.info('locals: ', locals.session.user);
 };
